Set video dates relative to timezone of file

+++ src/util/copyOrMoveSync.js
@@ -1,6 +1,21 @@
 import { copyFileSync, renameSync } from 'fs'
 
-export default function copyOrMoveSync({ copy, filename, newFilename }) {
+export default function copyOrMoveSync({
+  dryRun,
+  copy,
+  filename,
+  newFilename,
+}) {
+  if (newFilename === filename) {
+    return
+  }
+
+  console.log(`${filename.padEnd(70, '.')}${newFilename}`)
+
+  if (dryRun) {
+    return
+  }
+
   return copy
     ? copyFileSync(filename, newFilename)
     : renameSync(filename, newFilename)
